State

+++ src/car_state.rs
@@ -1,6 +1,8 @@
-use crate::game_logic::{CpuDifficulty, Orientation, Velocity};
+use crate::game_logic::{
+    apply_physics, calculate_steering_command, theta_star, CpuDifficulty, Orientation,
+    PhysicsInput, ThetaCheckpointList, ThetaCommand, ThetaGrid, Velocity,
+};
 use bevy::prelude::*;
-use rand::prelude::*;
 use std::time::Duration;
 
 // defining some states for our car
@@ -12,56 +14,20 @@ pub struct CarState {
     state: Option<Box<dyn State>>,
 }
 
-#[derive(Component)]
-struct FuseTime {
-    /// track when the bomb should explode (non-repeating timer)
-    timer: Timer,
-}
-
 // enums for the different transitions between car driving types
 enum Transition {
     None,
     ToAggressive,
     ToNeutral,
-    // ToAttack,
-}
-
-/**
-
-Random generator
-
-**/
-
-pub fn generate_number() -> i32 {
-    // Generates a value between 1 and 10
-    let mut rng = rand::rng();
-    return rng.random_range(1..=10);
 }
 
 impl CarState {
     pub fn new() -> CarState {
         CarState {
-            state: Some(Box::new(Neutral::new())), // default to aggressive
-        }
-    }
-    /*
-    pub fn to_defense(&mut self) {
-        if let Some(s) = self.state.take() {
-            self.state = Some(s.to_defense())
+            state: Some(Box::new(Neutral::new())),
         }
     }
 
-    pub fn to_aggressive(&mut self) {
-        if let Some(s) = self.state.take() {
-            self.state = Some(s.to_aggressive())
-        }
-    }
-
-
-    pub fn execute(&self) -> Transition {
-        self.state.as_ref().unwrap().execute(self)
-    }
-    */
     // this is what will be called every frame to control the behavior of the AI
     pub fn update(
         &mut self,
@@ -73,6 +39,8 @@ impl CarState {
         closest_car_position: Option<Vec2>,
         closest_car_distance: f32,
         difficulty: &CpuDifficulty,
+        checkpoints: &mut ThetaCheckpointList,
+        grid: &Res<ThetaGrid>,
     ) {
         if let Some(mut s) = self.state.take() {
             // do the current state's operations
@@ -86,6 +54,8 @@ impl CarState {
                 closest_car_position,
                 closest_car_distance,
                 difficulty,
+                checkpoints,
+                grid,
             );
 
             // transition based off of what each state returns
@@ -93,7 +63,6 @@ impl CarState {
                 Transition::None => s,
                 Transition::ToNeutral => s.to_neutral(),
                 Transition::ToAggressive => s.to_aggressive(),
-                // Transition::ToAttack => s.to_attack(),
             });
         }
     }
@@ -103,7 +72,6 @@ impl CarState {
 trait State: Send + Sync {
     fn to_neutral(self: Box<Self>) -> Box<dyn State>;
     fn to_aggressive(self: Box<Self>) -> Box<dyn State>;
-    // fn to_attack(self: Box<Self>) -> Box<dyn State>;
 
     // execute will return true in the case of a success
     // execute should contain some conditions to change to different states
@@ -117,11 +85,23 @@ trait State: Send + Sync {
         closest_car_position: Option<Vec2>,
         closest_car_distance: f32,
         difficulty: &CpuDifficulty,
+        checkpoints: &mut ThetaCheckpointList,
+        grid: &Res<ThetaGrid>,
     ) -> Transition;
 }
 
 // the state objects are aggressive, Neutral, etc.
-struct Aggressive {}
+struct Aggressive {
+    ram_timer: Timer,
+}
+
+impl Aggressive {
+    pub fn new() -> Self {
+        Self {
+            ram_timer: Timer::new(Duration::from_millis(2000), TimerMode::Once),
+        }
+    }
+}
 
 impl State for Aggressive {
     // TRANSITIONS BETWEEN STATES
@@ -130,7 +110,7 @@ impl State for Aggressive {
         Box::new(Neutral::new())
     }
     fn to_aggressive(self: Box<Self>) -> Box<dyn State> {
-        self
+        Box::new(Aggressive::new())
     }
     // --------------------------
     fn execute(
@@ -143,16 +123,63 @@ impl State for Aggressive {
         closest_car_position: Option<Vec2>,
         closest_car_distance: f32,
         difficulty: &CpuDifficulty,
+        _checkpoints: &mut ThetaCheckpointList,
+        _grid: &Res<ThetaGrid>,
     ) -> Transition {
+        self.ram_timer.tick(delta_time.delta());
+
+        if self.ram_timer.finished() {
+            return Transition::ToNeutral;
+        }
+
         // MAIN DRIVING LOGIC GOES HERE
         if car_nearby {
             if let Some(target_pos) = closest_car_position {
-                let ai_pos = transform.translation.truncate();
-                let direction = (target_pos - ai_pos).normalize_or_zero();
-                // info!("[+] AGGRESSIVE MODE: Going towards car at distance {:.1}!", closest_car_distance);
-                // info!("cpu difficulty: {:?}", difficulty);
-                // TODO: use theta* to pursue the target position
-                // TODO: increase velocity and adjust orientation to ram the target
+                let start_pos = (transform.translation.x, transform.translation.y);
+                let command = calculate_steering_command(
+                    start_pos,
+                    (target_pos.x, target_pos.y),
+                    orientation.angle,
+                );
+
+                let mut input = PhysicsInput::default();
+                match command {
+                    ThetaCommand::Forward => {
+                        input.forward = true;
+                    }
+                    ThetaCommand::Reverse => {
+                        input.backward = true;
+                    }
+                    ThetaCommand::TurnLeft => {
+                        input.left = true;
+                        input.forward = true;
+                    }
+                    ThetaCommand::TurnRight => {
+                        input.right = true;
+                        input.forward = true;
+                    }
+                    ThetaCommand::Stop => {
+                        // do nothing
+                    }
+                }
+
+                // For ramming, let's add boost.
+                input.boost = true;
+
+                let mut pos = transform.translation.truncate();
+                apply_physics(
+                    &mut pos,
+                    velocity,
+                    orientation,
+                    &input,
+                    delta_time.delta_secs(),
+                    1.0, // speed_modifier
+                    1.0, // friction_modifier
+                    1.0, // turn_modifier
+                    1.0, // decel_modifier
+                );
+                transform.translation.x = pos.x;
+                transform.translation.y = pos.y;
             }
         }
 
@@ -183,13 +210,11 @@ impl State for Neutral {
     // TRANSITIONS BETWEEN STATES
     // --------------------------
     fn to_neutral(self: Box<Self>) -> Box<dyn State> {
-        self // we're already in Neutral...
+        Box::new(Neutral::new())
     }
 
     fn to_aggressive(self: Box<Self>) -> Box<dyn State> {
-        Box::new(Aggressive {})
-        // use theta star to target player
-        // increase velocity to ram
+        Box::new(Aggressive::new())
     }
     // --------------------------
     fn execute(
@@ -202,15 +227,16 @@ impl State for Neutral {
         closest_car_position: Option<Vec2>,
         closest_car_distance: f32,
         difficulty: &CpuDifficulty,
+        checkpoints: &mut ThetaCheckpointList,
+        grid: &Res<ThetaGrid>,
     ) -> Transition {
         // MAIN DRIVING LOGIC GOES HERE
-        // TODO: use transform, velocity, etc to move the car
         self.decision_timer.tick(delta_time.delta());
 
         // check if a car is nearby - if so, immediately switch to aggressive
         if car_nearby {
             info!(
-                // "[+] car detected at distance {:.1}! Switching to aggressive mode!",
+                "[+] car detected at distance {:.1}! Switching to aggressive mode!",
                 closest_car_distance
             );
             return Transition::ToAggressive;
@@ -218,47 +244,44 @@ impl State for Neutral {
 
         // using theta* here to drive normally
         if self.decision_timer.just_finished() {
-            // info!("[+] Neutral driving - no cars nearby");
+            let start_pos = (transform.translation.x, transform.translation.y);
+            let command = theta_star(start_pos, orientation.angle, checkpoints, grid);
+
+            let mut input = PhysicsInput::default();
+            match command {
+                ThetaCommand::Forward => {
+                    input.forward = true;
+                }
+                ThetaCommand::Reverse => {
+                    input.backward = true;
+                }
+                ThetaCommand::TurnLeft => {
+                    input.left = true;
+                    input.forward = true;
+                }
+                ThetaCommand::TurnRight => {
+                    input.right = true;
+                    input.forward = true;
+                }
+                ThetaCommand::Stop => {
+                    // do nothing
+                }
+            }
+            let mut pos = transform.translation.truncate();
+            apply_physics(
+                &mut pos,
+                velocity,
+                orientation,
+                &input,
+                delta_time.delta_secs(),
+                1.0,
+                1.0,
+                1.0,
+                1.0,
+            );
+            transform.translation.x = pos.x;
+            transform.translation.y = pos.y;
         }
         Transition::None
     }
 }
-// struct Attack {}
-
-// impl State for Attack {
-//     // TRANSITIONS BETWEEN STATES
-//     // --------------------------
-//     fn to_attack(self: Box<Self>) -> Box<dyn State> {
-//         self // we're already in Neutral...
-//     }
-
-//     fn to_aggressive(self: Box<Self>) -> Box<dyn State> {
-//         Box::new(Aggressive {})
-//     }
-//     // --------------------------
-
-//     fn execute(&self,
-//         transform: &mut Transform,
-//         velocity: &mut Velocity,
-//         orientation: &mut Orientation,
-//     ) -> Transition {
-//         // MAIN DRIVING LOGIC GOES HERE
-//         // TODO: use transform, velocity, etc to move the car
-
-//         info!("Attacking!");
-//         info!("{:?}", transform.translation);
-
-//         info!("SPAWN MISSILE HERE!");
-
-//         let some_driving_condition: bool = true;
-//         if some_driving_condition == true {
-//             info!("Switching to offensive driving!");
-//             Transition::ToAggressive
-//         } else {
-//             Transition::None
-//         }
-//     }
-// }
-
-// // this is doing some weird rust ownership stuff I don't fully understand
-// // i just sort of copied the structure from the rust book and added extra bevy functions
